Rename tag collection and remove unused base classes.
