+++ lib/ibm/cloud/sdk/tags.rb
@@ -8,7 +8,7 @@ module IBM
   module Cloud
     module SDK
       # A collection interface into the Tags API.
-      class Tags < TAG::HTTP::BaseCollection
+      class Tags < TAG::TagCollection
         def initialize(connection, token, logger: nil) # rubocop:disable Lint/MissingSuper
           @token = token
           @connection = connection
