+++ lib/ibm/cloud/sdk/tags/http.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
 require 'ibm/cloud/sdk_http'
-require_relative 'http/base_mixin'
 
-require_relative 'http/base_collection'
-require_relative 'http/base_instance'
+require_relative 'http/tag_collection'
