--- lib/ibm/cloud/sdk/tags/http/base_instance.rb
@@ -1,17 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-require_relative '../http'
-
-module IBM
-  module Cloud
-    module SDK
-      module TAG
-        module HTTP
-          class BaseInstance < SDKHTTP::BaseInstance
-          end
-        end
-      end
-    end
-  end
-end
