--- lib/ibm/cloud/sdk/tags/http/base_mixin.rb
@@ -1,18 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-require_relative '../http'
-
-module IBM
-  module Cloud
-    module SDK
-      module TAG
-        module HTTP
-          module BaseMixin
-            include SDKHTTP::BaseHTTPMixin
-          end
-        end
-      end
-    end
-  end
-end
